# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FutureTeam API",
        "contact": {
            "email": "razraz.odinodin@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://futureteam.local/api/v1"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Логин через email",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\Auth\\LoginController::login",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "example": "test@test.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная авторизация, возвращает токен",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает массив ошибок",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Неправильный пароль или email",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выход из приложения",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\Auth\\LoginController::logout",
                "responses": {
                    "204": {
                        "description": "Удаляет все токены клиента",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register by email",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\Auth\\RegisterController::register",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "unique email",
                        "required": true,
                        "example": "test@test.com"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "'company' or 'member'",
                        "required": true,
                        "example": "member"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешная регистрация, возвращает сущность User"
                    },
                    "400": {
                        "description": "Возвращает массив ошибок"
                    }
                }
            }
        },
        "/common/activityfields": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Сферы деятельности",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\Common\\ActivityFieldsController::index",
                "responses": {
                    "200": {
                        "description": "Возвращает список сфер деятельности",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/common/goals": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Цели устойчивого развития ООН",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\Common\\GoalsController::index",
                "responses": {
                    "200": {
                        "description": "Возвращает список целей",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/common/languages": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Языки",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\Common\\LanguagesController::index",
                "responses": {
                    "200": {
                        "description": "Возвращает список языков",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/common/countries": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Страны",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\Common\\PlacesController::countries",
                "responses": {
                    "200": {
                        "description": "Возвращает список стран",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/common/cities": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Города",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\Common\\PlacesController::cities",
                "parameters": [
                    {
                        "name": "country_id",
                        "in": "query",
                        "description": "Id страны",
                        "required": true
                    },
                    {
                        "name": "value",
                        "in": "query",
                        "description": "Начало ввода города(мин длинна 2 символа))",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает список городов",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает массив ошибок",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/common/city/{city}": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Город",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\Common\\PlacesController::city",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "description": "Id города",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает город по его id",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/": {
            "get": {
                "tags": [
                    "Api base"
                ],
                "summary": "Base path",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\HomeController::index",
                "responses": {
                    "200": {
                        "description": "Api information",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/account": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Возвращает профиль текущего пользователя",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\User\\UsersController::account",
                "responses": {
                    "200": {
                        "description": "Возвращает профиль текущего пользователя",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Добавляет профиль к пользователю",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\User\\UsersController::setProfile",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "Имя",
                                        "type": "string"
                                    },
                                    "middle_name": {
                                        "description": "Отчество",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "Фамилия",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Телефон",
                                        "type": "string"
                                    },
                                    "birth_date_at": {
                                        "description": "Дата рождения",
                                        "type": "string"
                                    },
                                    "city_id": {
                                        "description": "ID города",
                                        "type": "integer"
                                    },
                                    "activity_field_id": {
                                        "description": "ID Сферы деятельности",
                                        "type": "integer"
                                    },
                                    "goals[0]": {
                                        "description": "ID Цели",
                                        "type": "integer"
                                    },
                                    "goals[1]": {
                                        "description": "ID Цели",
                                        "type": "integer"
                                    },
                                    "known_languages[]": {
                                        "description": "Языки, которыми владею(массив Id)",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "collectionFormat": "multi"
                                    },
                                    "languages_wltl[]": {
                                        "description": "Языки, которые хочу выучить(массив Id)",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "collectionFormat": "multi"
                                    },
                                    "language_exchange_agreement": {
                                        "description": "Согласие на участие в языковом обмене (0 или 1)",
                                        "type": "integer"
                                    },
                                    "motivation[]": {
                                        "description": "Цели(массив строк, длинна 1-5)",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "collectionFormat": "multi"
                                    },
                                    "photo": {
                                        "description": "Фото",
                                        "type": "file",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "about": {
                                        "description": "О себе",
                                        "type": "string"
                                    },
                                    "terms": {
                                        "description": "Согласие на обработку данных"
                                    },
                                    "organization_id": {
                                        "description": "id существующей организации",
                                        "type": "integer"
                                    },
                                    "organization_name": {
                                        "description": "Имя организации",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешное добавление профиля к пользователю",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает массив ошибок",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "parameters": {
            "token": {
                "name": "Authorization",
                "in": "header",
                "required": true,
                "example": "Bearer "
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Api base",
            "description": "API base page"
        },
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "User",
            "description": "Пользователи"
        },
        {
            "name": "Common",
            "description": "Общая информация"
        }
    ]
}